Don't supply ArgumentConvert with empty string if args is empty
Usually, implementations of PopArgument (that's the trait in poise that is implemented for all valid parameter types) return a dedicated error if the input is empty. But the blanket impl for serenity::ArgumentConvert used Err = T::Err for simplicity. That means there was no room for a dedicated error: we used T's Err directly. I changed it to Err = WrapperArgumentParseError<T::Err> where WrapperArgumentParseError is an enum that's either EmptyArgs (and it reuses string's EmptyArgs ZST error struct), or T::Err.