Use String instead of impl ToString again
The entire library currently tends to avoid "convenience" generics wherever possible. I'm not 100% sure that's the right call, but in any case the guideline should apply to the entire library, or not at all. So I'm removing this instance of impl ToString in favor of consistency with the rest of the library as of now