Add common trait implementations to many types
Not a lot of types, and not a lot of traits either, because most structures in poise are a giant tree structure where at some point in the hierarchy there's a stubborn type that's incompatible with common traits

Most commonly: fn pointers. They SUPPOSEDLY implement a lot of common traits, but in practice all of the trait implementations cease to work (afaict) when there are any lifetimes involved at all. So, no PartialEq or Debug for a lot of these types

Another barrier is RwLock and Mutex