Don't apply reuse_response for slash
Why? It all begins with modals. When you send a modal as the initial response (it has to be), and the command is marked reuse_response, poise was trying to edit the modal as if it were a message and crash. This didin't ever fail before because pre-modals, all initial interaction response possibilities were editable into text. Anyways, there's a bunch of ugly solutions, like keeping track in ApplicationContext not only whether a response was sent, but also whether it was a modal, in which case the reuse_response editing would be directed at a followup response instead, so you'd also keep track of the followup response message ID and it's all a big mess. And then I remembered why reuse_response exists in the first place and it's for prefix commands, for edit tracking. And using reuse_response for slash doesn't even really make sense, like, think about it, you're sending multiple messages in the code (`ctx.send(1); ctx.send(2);`) but it all goes in a single message. Makes no sense. If people wanna edit their initial response, they should do _that_: `ctx.send(1).edit(2);`. And this now also works absolutely fine with modals. You can do `Modal::execute(ctx); ctx.send(1).edit(2);` and the edit will not try to edit the modal, but it will edit the send(1) which is already a followup message because Modal::execute set the has_sent_initial_response atomic flag, and everything was fine. I hope. We will see