Don't be as quick to trigger cooldown
Previously, the cooldown trigger was done directly after checking it. Now, the trigger is delayed, so:
- autocomplete callbacks doesn't trigger the cooldown anymore (still respect it though)
- argument parse errors don't trigger cooldown anymore

Though, now we have to wait until a command finishes execution before trigger cooldown, because it might return Err(FrameworkError::ArgumentParse). So, users can now bypass the cooldown when invoking a command again before the first invocation finishes. But anden3 says this is ok so :shrug: https://discord.com/channels/381880193251409931/919310428344029265/968599843968806962