Introduce FrameworkContext in favor of &Framework
Basically, instead of passing the full blown Framework around, have a small POD type that contains the data users care about that they used to get from Framework directly. This is good because it opens the possibility of running the dispatch functions outside of a poise::Framework. I.e., it allows users to have their own event loop. In principle. This is an MVP commit