Fix CreateReply::to_prefix_edit
This was a nuanced issue. Previously, to_prefix_edit reset fields that were not explicitly edited. I did that ages ago for edit tracking; writing a message into an existing message shouldn't keep anything of the existing message. However, now, to_prefix_edit is used not just for edit tracking but also for ReplyHandle::edit(). And there, this behavior is wrong. I noticed it in register_application_commands_buttons where it calls `.edit(|b| b.components(|b| b))` to remove the components from the message. However, it made the whole message blank because of the behavior.

In this commit, I changed to_prefix_edit to not reset anything. The resetting is done manually in the edit tracking code. This behavior makes much more sense, because all the other to_* methods on CreateReply don't reset stuff either