Prevent underscore binding lint (#216)
This lint is triggered when we have an unused parameter in the handler function. The code that's generated copies the identifier name from the function definition, but does use the identifier, so we need to disable this lint. That said, a better solution might be to add a fixed prefix like `poise_arg_` to each parameter name.