Remove specialization hacks for `PopArgument` and `SlashArgument` (#357)
Replaces the specialization of blanket impls with macro-generated impls for
types that explicitly implement `serenity::ArgumentConvert`, and introduce a
new `#[string]` attribute to signal that a type's `FromStr` implementation
should be used.

Also replaces hacky blanket impls with macro-generated impls for specific
types, including some types that did not have impls before, to line up with the
list of types that implement `PopArgument`.