feat!: Add burstable counts to cooldowns
This enables for commands to be invoked multiple times before they get
put on cooldown. A cooldown cycle begins with the first instance a
command is invoked and ends when the cooldown expires. When the cooldown
expires, the number of invokes also resets to 0.

By default, the burstable amount for each type of cooldown is set to 1,
which will mimic the old behavior, only allowing for 1 command invoke
per cooldown cycle. Command attribute macros that do not explicitly add
these fields will inherit the default value.

The reason why this change is marked as breaking is due to
`CooldownConfig` having additional public fields being added. This
struct was already marked as non-exhaustive, so the user should expect
this struct to break.